added parralel coloring correctly
