+++ .gitignore
@@ -1,8 +1,3 @@
 .vscode/
 .spyproject/
 debug/
-*.laz
-*.las
-*.sh
-point-cloud-colorize/*
-get_year.sql
