--- pipeline.json
@@ -1,9 +1,0 @@
-{
-  "pipeline": [
-    "/var/data/arnot/div_color/tmp_col/*.laz",
-    {
-        "type" : "filters.merge"
-    },
-   "/var/data/arnot/div_color/c_10cn2_color.laz"
-  ]
-}
