+++ point_cloud_colorize/__init__.py
@@ -1,0 +1,1 @@
+name = 'point_cloud_colorize'
